Reem A Homework1
How difficult did you find this (out of 10)? 0 being no problems at all, 10 being impossible;
3
Was there anything that you struggled with?;
undrstanding the qoustion and it was alot
Is there anything that you’d like some further information on?;
i like homework but it was alot so plz not that much next time :)
Roughly how long did it take?
4 to 5 h